[ios] Removes Infobar banner drag down gesture.
Removes the ability to swipe down the Infobar banner in order to reveal
the modal. Now only tapping on the Banner should display a Modal.

This is part of the change UX are proposing after the experiment.

Bug: 961343
Change-Id: Ia64198b22ff9c7d32c129f7532ba4491ebb063a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804081
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#696866}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7c5eebcafbd53ff679fbcc06e446918832437a7