[ios] Create Overflow badge
This change adds an overflow badge. Since it is the first
non-infobar badge, BadgeTappableItem is created to configure
it. BadgeLocalDispatcher will send a signal to the LocationBarCoordinator
in response to tapping the badge. Ultimately LocationBarCoordinator
will relay that signal to BrowserViewController with badge model information
to present the overflow popup menu.

Since badges should not be shown both in the BadgeViewController and the
popup menu, this will allow the BadgeViewController to dynamically decide what badges
it will show and then pass the rest to the popup menu. While that could
be possible through matching assumptions in the separate popupmenu and
badge mediator, it is possible that the BadgeViewController will decide on how
many badges it shows based on available space.

Bug: 976901
Change-Id: If3d977ae49949c0d65413bb61d6028e5febb2605
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1789852
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#696918}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b2e17f01cc10479e46607e88958b46609873753