[ios] Remove Unused FullScreen Feature Flags
There is no need for ViewportAdjustmentExperiment since the only experiment
remaining is just the kSmoothScrollingDefault feature flag.
GetActiveViewportExperiment() can just be a function that returns
whether or not smooth scrolling is turned on. GetActiveViewportExperiment()
and ViewportAdjustmentExperiment are still needed for downstream compatibility,
and will be removed in a subsequent CL.

Bug: 914042
Change-Id: I57bcd13dafaafb3b3a05ec65f23dacf2936e81e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1799663
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#697673}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9384d0a257a92133f77b787847bb8466965045cc