[ios/web] Don't associate nil wk_item's.
Due to an unrelated issue, it's possible for about:// urls to be
written as an empty url.  Because an empty forward or backwards item
will return an empty URL, it was possible to associate to a nil
WKBackForwardItem and crash.

Bug: 1005211
Change-Id: I2becd93a8d35cde0f1a68fc2ac3da6853c023392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1810803
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#697892}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5444c345c8e22ba2f845c2facfcead34495f091d