[ios] Adds badgeSupport parameter to InfobarCoordinator initializer.
Some Infobar Messages might not add/have a badge. This CL adds a parameter
to the initializer method to add support for this.

The InfobarCoordinator BadgeDelegate will not be set if
the Coordinator shouldn't support a badge.

Bug: 961343
Change-Id: I094b9ae720578d12f6c55e57c66fdc98dbfe1bc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1811840
Reviewed-by: Peter Lee <pkl@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#698232}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 533b24b763cb41c901a3ca2676b0f06d34da56bc