[iOS] Do not store file:// URL in history
From my tests, it seems that the file:// URL can only access files that
are stored in the app folder. For now the only folder we have is emptied
every time the app is launched, so storing the URL in the history don't
make sense as the user won't be able to revisit them.

Bug: 1003234
Change-Id: I7b5795009b0f5ddb2fb68f4eb4df476ddfe89b8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1803148
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#699279}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5466161bad62960f572d854734290a1273b68408