[Passwords] Remove AutofillAgent::ShowInitialPasswordAccountSuggestions
This was to support the fill-on-select experiment.
The current UI is very hard and will mostly not going launch.
This CL is replacing the behavior of opening of prompts on page load
with existing Fill on account select behavior (which is already working
 on PSL matching and http forms).

Using the existing behavior has multiple benefits:
1.decrease maintenance cost
2.ensure that this feature works
3.collects UMA on using this feature

TBR=mkwst@chromium.org,vasilii@chromium.org

Change-Id: I7a0b19dd18a56ae86712b34429023c96ceab9404
Bug: 949519
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1821524
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#699688}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 766a4b0360e6d2cc141c2b4db3167f99928274b4