[iOS] Early return if the navigation context is nil during error
When failing a provisional navigation, the context might be nil
(probably a race condition in the callbacks of the WKWebView).
This issue was handled in handleCancelledError:forNavigation:
provisionalLoad: but wasn't in the other callbacks.
This CL makes sure that we handle it.
The workaround in crrev.com/c/1700054 was probably preventing the
crash from happening.

Bug: 1004634
Change-Id: I1f94812617c908a63124c16a18a3ca71d9902d8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1816575
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#699741}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 517ce639e922d96844664657d3321090579ef286