[iOS] Change ownership of WebStateList.
Currently, TabModels create and own their WebStateLists. A
Browser creates a TabModel, and then exposes its WebStateList.

There are two problems with this. First, TabModel is being
actively refactored away. Second, because TabModel is an
Objective-C object, this makes it hard to reason about when
the WebStateList will be destroyed, which makes things like
dangling observers hard to catch (see crbog.com/1001960, for
example).

This CL changes this ownership. Now, BrowserImpl creates and
owns the WebStateList, and it passes a (weak) pointer into
TabModel. TabModel's existing -browserStateDestroyed method is
called when the Browser destructs, and the contract of that
method is that the TabModel should tear down its interactions
with the Browser and BrowserState, and it (the TabModel) should
be removed soon thereafter.

BrowserImpl now also creates and owns the web state list
delegate; its current role is just to attach tab helpers each
time a web state is added to the list. This used to be named
TabModelWebStateListDelegate, but this CL renames and moves it
to be BrowserWebStateListDelegate.

Various tests need to now create web state lists (and, thus,
web state list delegates) to inject into tab models.

-browserStateDestroyed is renamed to -disconnect to match other
code.

Because a test in ios_internal/ uses the existing TabModel
initializer, it needs to be kept functional (which involves it
creating an internal web state list). This will go away as
soon as possible, once this CL lands and the internal test can
be cleaned up.

Bug: 1001960
Change-Id: I8d8886bbbedc4c701ae0956d0f5e0984812b725b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1796358
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#700676}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e60338bba9b8125fcbbcff9f35d6c3b86e9139ba