[ios/web] Fix crash related to kRestoreNavigationTime fix.
It's possible for there to be pending navigations for a session that is
going to be restored (such as for the -ForwardHistoryClobber workaround).
In this case, the pending navigation will start while the navigation
manager is in restore mode.  There are other edges cases where a restore
session finishes without trigger it's start, such as when restoring some
with some app specific or blocked URLs, or when WKWebView's
backForwardList state is out of sync.

Bug: 1008026
Change-Id: I2d4d41321a21f8dad8acf72c8e8529ac73ea7d03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826890
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Auto-Submit: Justin Cohen <justincohen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#700677}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7f334b670f6db360efe5a95ab81bfce27d1afb5