Check last committed item when loading data.
Crash reports show that the last committed item can be null.
It is not clear at the moment what steps can be to this scenario, but
this CL adds a check to avoid crashing when this happens.

Bug: 1005210
Change-Id: Ib29b0ffb5e8ca09821c200e1765ba269846ea171
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1827283
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#701082}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b07ee65277f19934eae6fdf8b84208bddff61aa2