[ios] Add BadgeStatePresenting BadgeState
This state will be important when a badge's banner is presented
and there are multiple non-fullscreen badges. In this situation,
the badge will be displayed while the banner is presented. When
the banner dismisses, the badge will be replaced by the overflow
badge.

Bug: 976901
Change-Id: Ic8d8b7571a468edd7bd0d7e18ec046fe4ce462ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825868
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#701325}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80c470b92cc16741fcfee48ec31acc3b788b76b7