[iOS] Update tools menu VoiceOver
This CL removes the default VoiceOver action which is reading the title
of the item then the content of the badge.
Now the accessibility label is the title and a new property,
|additionalAccessibilityLabel|, if it is set.

This is a reland of crrev.com/c/1789397, fixing the failing tests.

Bug: 985764
Change-Id: I6abef0554fd30463d3381de7baf000752cafeb39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1833475
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#702782}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ed5a7233af3d71138735b8a01a3673a34e6ed9c